// PFF-CRMEF-Client/src/app/model/stagiaire.ts
export class Stagiaire {

    idStag: number;
    cne: string;
    cin: string;
    nomPrenom: string;
    dateNais: Date;
    lieuNais: string;
    adresse: string;
    tele: string;
    email: string;
    picByte: string;
    sexe: string;
    specialite: string;
    corpsEducatif: string;
    handicap: string;
    stuation: string;
    etat: string;




    //Bac
    serieBac: string;
    mentionBac: string;
    anneeObtenu: Date;
    provinceObtenu: string;
    noteObtenuBac: String;
    academie: string;


    // licence 
    codeApoge: string;
    dernierDiplome: string;
    specialiteDiplome: string;
    noteObtenuDiplome: string;
    mentionDiplome: string;
    faculterObtenuDiplome: string;
    anneeDiplome: Date;

    //parent
    nomPere: string;
    nomMere: string;
    profesionPere: string;
    profesionMere: string;
    codePostale: string;
    teleParent: string;
    teleFixParent: string;
    adresseParents: string;



}